Revert "Support Wide String in FBX"
This reverts commit a86cb9dcb0624d490500424c7521115325813152.
